requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.10
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
